Add new project dock with curated content
Perform major surgery to the code, destroy much in the process. I guess this one wont be salvagable. It probably need to be done all over again. It try to split the view and XMI into separated entities, but it is half assed to say the least.
That said, it does actually work!